+++ src/components/ProfileCard/ProfileCard.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { ButtonGroup, Dropdown } from 'react-bootstrap';
 import { useNavigate, useParams } from 'react-router-dom';
 import useLocalStorage from 'utils/useLocalstorage';
-import styles from './../../style/app.module.css';
+import styles from './../../style/app-fixed.module.css';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 
 /**
