+++ src/style/app.module.css
@@ -607,18 +607,6 @@
   border: 1px solid var(--dropdown-border);
 }
 
-.dropdown:not(:focus) {
-  /* background-color: var(--bs-white);
-  border: 1px solid var(--brown-color);
-  color: var(--brown-color) !important; */
-}
-
-/* .dropdown:disabled {
-  background-color: transparent !important;
-  border: 1px solid var(--brown-color);
-  color: var(--brown-color) !important;
-} */
-
 .dropdownItem {
   background-color: var(--dropdown-item-bg) !important;
   color: var(--dropdown-item-text) !important;
@@ -8853,6 +8841,7 @@ button[data-testid='createPostBtn'] {
 }
 
 /* ProfileCard.tsx */
+
 .profileContainer {
   border: none;
   padding: 2.1rem 0.5rem;
